+++ server/src/service/applications.js
@@ -40,8 +40,16 @@ const images = {
   },
 };
 
+const buildUrlOfDetail = (aPublicUrl, aCompanyType, aJobId) => {
+  let res = "";
+  if (aCompanyType === "matcha") {
+    res = `${aPublicUrl}/recherche-apprentissage?display=list&page=fiche&type=${aCompanyType}&itemId=${aJobId}`;
+  }
+  return res;
+};
+
 const initApplication = (query, companyEmail) => {
-  return new Application({
+  let res = new Application({
     applicant_file_name: query.applicant_file_name,
     applicant_email: query.applicant_email.toLowerCase(),
     applicant_first_name: query.applicant_first_name,
@@ -58,6 +66,8 @@ const initApplication = (query, companyEmail) => {
     job_id: query.job_id,
     interet_offres_mandataire: query.interet_offres_mandataire,
   });
+
+  return res;
 };
 
 const getApplications = async (qs) => {
@@ -146,6 +156,16 @@ const sendApplication = async ({ mailer, query, shouldCheckSecret }) => {
 
       const fileContent = query.applicant_file_content;
 
+      const urlOfDetail = buildUrlOfDetail(publicUrl, query.company_type, query.job_id);
+
+      const buildTopic = (aCompanyType, aJobTitle) => {
+        let res = "contact";
+        if (aCompanyType === "matcha") {
+          res = `Candidature en alternance - ${aJobTitle}`;
+        }
+        return res;
+      };
+
       // Sends acknowledge email to "candidate" and application email to "company"
       const [emailCandidat, emailCompany] = await Promise.all([
         mailer.sendEmail(
@@ -162,9 +182,9 @@ const sendApplication = async ({ mailer, query, shouldCheckSecret }) => {
         ),
         mailer.sendEmail(
           application.company_email,
-          `Candidature spontanée pour un poste en alternance`,
+          buildTopic(application.company_type, application.job_title),
           getEmailTemplate(emailTemplates.entreprise),
-          { ...application._doc, ...images, ...encryptedId, publicUrl },
+          { ...application._doc, ...images, ...encryptedId, publicUrl, urlOfDetail },
           [
             {
               filename: application.applicant_file_name,