+++ server/CHANGELOG.md
@@ -5,6 +5,7 @@
 [LBA6-131] Scrolling vers les offres cassé
 [LBA6-135] Restauration info métier Matcha
 [LBA6-11] MAJ email envoyé au recruteur lors d'une candidature Matcha
+[LBA6-12] MAJ email envoyé au recruteur lors d'une candidature spontanée
 
 ## 1.18.11 [11/05/2022]
 [LBA6-32] Contrôle des emails des candidats : pas d'email temporaire